Switch sign of PES error
The PES rule previously moved in the direction of what it called the
"error" signal, meaning that the error signal had to be the target
value for the learning minus the actual value. This is the opposite
of what people typically refer to as error. This commit changes the
sign of this error so that it is actually an error, and the PES rule
moves in the opposite direction of the error.